create setup_filebeat.sh, update all scripts

+++ setup_elk.sh
@@ -4,13 +4,34 @@
 
 set -e  # Прерываем скрипт при ошибке
 
+echo "Проверка установки elasticsearch"
+if ! dpkg -s elasticsearch &>/dev/null; then
+  echo "Пакет elasticsearch не установлен. Установите его через: sudo apt install elasticsearch"
+  exit 1
+fi
+
+echo "Проверка установки logstash"
+if ! dpkg -s logstash &>/dev/null; then
+  echo "Пакет logstash не установлен. Установите его через: sudo apt install logstash"
+  exit 1
+fi
+
+echo "Проверка установки kibana"
+if ! dpkg -s kibana &>/dev/null; then
+  echo "Пакет kibana не установлен. Установите его через: sudo apt install kibana"
+  exit 1
+fi
+
 # Определим IP-адрес хоста
 # HOST_IP=$(ip -o -4 addr show scope global | awk '{print $4}' | cut -d/ -f1 | head -n1)
 
-### === 1. Elasticsearch ===
-echo "[+] Настройка Elasticsearch..."
+### Elasticsearch
+echo "Настройка Elasticsearch"
+# Создаем резервную копию оригинального файла
+cp /etc/elasticsearch/elasticsearch.yml /etc/elasticsearch/elasticsearch.yml.bak
+
 sudo tee /etc/elasticsearch/elasticsearch.yml > /dev/null <<EOF
-network.host: 0.0.0.0
+http.host: 0.0.0.0
 http.port: 9200
 
 cluster.name: elk-cluster
@@ -24,10 +45,12 @@ path.logs: /var/log/elasticsearch
 xpack.security.enabled: false
 xpack.security.transport.ssl.enabled: false
 xpack.security.http.ssl.enabled: false
+
+
 EOF
 
-### === JVM Options ===
-echo "[+] Установка лимита памяти для Elasticsearch (1 ГБ)..."
+### JVM Options
+echo "Установка лимита памяти для Elasticsearch (1 ГБ)"
 sudo mkdir -p /etc/elasticsearch/jvm.options.d
 
 sudo tee /etc/elasticsearch/jvm.options.d/jvm.options > /dev/null <<EOF
@@ -42,18 +65,34 @@ sleep 5
 
 # Проверка Elasticsearch
 if curl -s http://localhost:9200 >/dev/null; then
-  # echo "[✓] Elasticsearch работает на http://$HOST_IP:9200" 
-  echo "[✓] Elasticsearch работает на http://192.168.56.107:9200"
+  # echo "Elasticsearch работает на http://$HOST_IP:9200" 
+  echo "[OK] Elasticsearch работает на http://192.168.56.107:9200"
 else
   echo "[!] Elasticsearch не отвечает. Проверьте журнал: journalctl -u elasticsearch"
 fi
 
-### === 2. Kibana ===
-echo "[+] Настройка Kibana..."
+### Kibana
+
+echo "Настройка Kibana"
+# Создаем резервную копию оригинального файла
+cp /etc/kibana/kibana.yml /etc/kibana/kibana.yml.bak
+
 sudo tee /etc/kibana/kibana.yml > /dev/null <<EOF
 server.port: 5601
 server.host: "0.0.0.0"
-elasticsearch.hosts: ["http://localhost:9200"]
+# elasticsearch.hosts: ["http://localhost:9200"]
+pid.file: /run/kibana/kibana.pid
+logging:
+  appenders:
+    file:
+      type: file
+      fileName: /var/log/kibana/kibana.log
+      layout:
+        type: json
+  root:
+    appenders:
+      - default
+      - file
 EOF
 
 sudo systemctl enable kibana
@@ -61,14 +100,17 @@ sudo systemctl restart kibana
 sleep 5
 
 if curl -s http://localhost:5601 >/dev/null; then
-  # echo "[✓] Kibana работает на http://$HOST_IP:5601"
-  echo "[✓] Kibana работает на http://192.168.56.107:5601"
+  # echo "[OK] Kibana работает на http://$HOST_IP:5601"
+  echo "[OK] Kibana работает на http://192.168.56.107:5601"
 else
   echo "[!] Kibana не отвечает. Проверьте журнал: journalctl -u kibana"
 fi
 
-### === 3. Logstash ===
-echo "[+] Настройка Logstash..."
+### Logstash
+echo "Настройка Logstash"
+# Создаем резервную копию оригинального файла
+cp /etc/logstash/logstash.yml /etc/logstash/logstash.yml.bak
+
 sudo mkdir -p /etc/logstash/conf.d
 
 sudo tee /etc/logstash/conf.d/logstash-nginx-es.conf > /dev/null <<EOF
@@ -107,16 +149,22 @@ output {
 }
 EOF
 
+sudo tee /etc/logstash/logstash.yml > /dev/null <<EOF
+
+path.data: /var/lib/logstash
+path.config: /etc/logstash/conf.d
+path.logs: /var/log/logstash
+EOF
+
 sudo systemctl enable logstash
 sudo systemctl restart logstash
 
 # Проверка состояния Logstash
 if systemctl is-active --quiet logstash; then
-  echo "[✓] Logstash настроен и работает. Слушает порт 5400 для логов от Filebeat"
+  echo "[OK] Logstash настроен и работает. Слушает порт 5400 для логов от Filebeat"
 else
   echo "[!] Logstash не запущен. Проверьте журнал: journalctl -u logstash"
 fi
 
-# Финальное сообщение
-# echo "[✓] ELK-стек настроен. Перейдите в Kibana: http://$HOST_IP:5601"
-echo "[✓] ELK-стек настроен. Перейдите в Kibana: http://192.168.56.107:5601"
+# echo "[OK] ELK-стек настроен. Перейдите в Kibana: http://$HOST_IP:5601"
+echo "[OK] ELK-стек настроен. Перейдите в Kibana: http://192.168.56.107:5601"
