+++ setup_prometheus.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 set -e  # Прерываем выполнение при ошибке
 
-echo "[INFO] Установка и настройка Prometheus..."
+echo "Проверка установки prometheus"
+if ! dpkg -s prometheus &>/dev/null; then
+  echo "Пакет prometheus не установлен. Установите его через: sudo apt install prometheus"
+  exit 1
+fi
+
+echo "Установка и настройка Prometheus"
+
+# Создаем резервную копию оригинального файла
+cp /etc/prometheus/prometheus.yml /etc/prometheus/prometheus.yml.bak
 
 # Создаём базовую конфигурацию Prometheus
 sudo tee /etc/prometheus/prometheus.yml > /dev/null <<EOF
@@ -18,13 +27,14 @@ scrape_configs:
           - '192.168.56.104:9100'  # apache2
 EOF
 
-echo "[INFO] Перезапуск Prometheus с новой конфигурацией..."
+echo "Перезапуск Prometheus с новой конфигурацией"
 sudo systemctl restart prometheus
 sudo systemctl enable prometheus
 
 # Определение IP-адреса для справочного вывода
-HOST_IP=$(ip -o -4 addr show scope global | awk '{print $4}' | cut -d/ -f1 | head -n1)
+# HOST_IP=$(ip -o -4 addr show scope global | awk '{print $4}' | cut -d/ -f1 | head -n1)
 
 echo "[OK] Prometheus настроен и запущен"
-echo "Интерфейс доступен по адресу: http://$HOST_IP:9090"
+# echo "Интерфейс доступен по адресу: http://$HOST_IP:9090"
+echo "Интерфейс доступен по адресу: http://192.168.56.106:9090"
 echo "Проверьте статус: sudo systemctl status prometheus"
